requirements.txt
numpy
pandas
matplotlib
seaborn
torch
torchvision
opencv-python
scikit-learn
imgaug
pillow
scipy
imageio
pydicom
tqdm
